refactor(NODE-3688): new tests to ts and async
